refactor: remove dead code and de-duplicate TLS/CSV paths (ponytail audit)
Whole-repo over-engineering audit. Removes verified-dead internal code and
collapses copy-pasted logic. No public-API changes: documented/exported
surface (ArrowConverter, ResultSetIterator, blocking_* sync API,
ArrowToCsvWriter, CsvToArrowReader, public error variants, SessionConfig
fields) was intentionally left in place.

Removed (zero non-test callers, not re-exported):
- error: AdbcErrorCode enum + Display + to_adbc_code() (unused; FFI uses adbc_core::Status)
- connection/session: SessionManager + dead Session methods/fields (get/set/remove_attribute, idle/timeout helpers, mark_error)
- transport/native/handshake: parse_login_response + count_attributes + attribute_wire_size
- transport/native/constants: ~30 never-referenced CMD_*/ATTR_*/COL_*/T_*/TRANSACTION_* constants
- export: write_arrow_ipc stream variant, build_schema_from_exasol_types, export_to_parquet_async_stream, unused ExportError variants
- query: Statement::bind_all skipped (public); removed compile-only export tests, get_file_name dup

De-duplicated:
- TLS verifiers (NoVerifier/FingerprintVerifier/verify_schemes) hoisted from 3 copies into src/transport/tls.rs
- export CSV: three hand-rolled parsers unified into one parse_csv_row; gzip/bzip2 decompress extracted
- import csv: is_compressed_file expressed via detect_compression

Verified against live Exasol Docker (2025.2.1): unit 1011 ok, integration
63 ok (2 pre-existing failures identical on main), import/export 42 ok,
driver-manager 41 ok. fmt + clippy (--all-targets --all-features -D warnings)
clean under pinned 1.92.0.
