+++ src/connection/session.rs
@@ -2,7 +2,7 @@
 //!
 //! This module handles session lifecycle, state tracking, and connection pooling support.
 
-use crate::connection::auth::{AuthResponseData, AuthenticationHandler};
+use crate::connection::auth::AuthResponseData;
 use crate::connection::version::parse_release_version;
 use crate::error::ConnectionError;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
@@ -124,9 +124,6 @@ pub struct Session {
     /// Current schema
     current_schema: Arc<RwLock<Option<String>>>,
 
-    /// Session attributes
-    attributes: Arc<RwLock<std::collections::HashMap<String, String>>>,
-
     /// Memoized native-Parquet-import capability, derived from `server_info.release_version`.
     native_parquet_cache: OnceLock<bool>,
 }
@@ -143,7 +140,6 @@ impl Session {
             query_count: AtomicU64::new(0),
             in_transaction: AtomicBool::new(false),
             current_schema: Arc::new(RwLock::new(None)),
-            attributes: Arc::new(RwLock::new(std::collections::HashMap::new())),
             native_parquet_cache: OnceLock::new(),
         }
     }
@@ -190,17 +186,6 @@ impl Session {
     pub async fn update_activity(&self) {
         let mut last_activity = self.last_activity.write().await;
         *last_activity = Instant::now();
-    }
-
-    /// Get time since last activity.
-    pub async fn idle_duration(&self) -> Duration {
-        let last_activity = self.last_activity.read().await;
-        last_activity.elapsed()
-    }
-
-    /// Check if session is idle beyond timeout.
-    pub async fn is_idle_timeout(&self) -> bool {
-        self.idle_duration().await > self.config.idle_timeout
     }
 
     /// Increment query counter.
@@ -277,24 +262,6 @@ impl Session {
         self.update_activity().await;
     }
 
-    /// Get a session attribute.
-    pub async fn get_attribute(&self, key: &str) -> Option<String> {
-        let attributes = self.attributes.read().await;
-        attributes.get(key).cloned()
-    }
-
-    /// Set a session attribute.
-    pub async fn set_attribute(&self, key: String, value: String) {
-        let mut attributes = self.attributes.write().await;
-        attributes.insert(key, value);
-    }
-
-    /// Remove a session attribute.
-    pub async fn remove_attribute(&self, key: &str) -> Option<String> {
-        let mut attributes = self.attributes.write().await;
-        attributes.remove(key)
-    }
-
     /// Close the session.
     pub async fn close(&self) -> Result<(), ConnectionError> {
         self.set_state(SessionState::Closing).await;
@@ -313,11 +280,6 @@ impl Session {
     /// Check if session is closed.
     pub async fn is_closed(&self) -> bool {
         matches!(self.state().await, SessionState::Closed)
-    }
-
-    /// Mark session as having an error.
-    pub async fn mark_error(&self) {
-        self.set_state(SessionState::Error).await;
     }
 
     /// Validate session is ready for operations.
@@ -351,103 +313,10 @@ impl std::fmt::Debug for Session {
     }
 }
 
-/// Session manager for connection pooling and lifecycle management.
-pub struct SessionManager {
-    /// Active sessions
-    sessions: Arc<RwLock<std::collections::HashMap<String, Arc<Session>>>>,
-
-    /// Session factory (authentication handler)
-    _auth_handler: Arc<AuthenticationHandler>,
-
-    /// Session configuration
-    config: SessionConfig,
-}
-
-impl SessionManager {
-    /// Create a new session manager.
-    pub fn new(auth_handler: Arc<AuthenticationHandler>, config: SessionConfig) -> Self {
-        Self {
-            sessions: Arc::new(RwLock::new(std::collections::HashMap::new())),
-            _auth_handler: auth_handler,
-            config,
-        }
-    }
-
-    /// Register a new session.
-    pub async fn register_session(&self, session: Arc<Session>) {
-        let mut sessions = self.sessions.write().await;
-        sessions.insert(session.session_id().to_string(), session);
-    }
-
-    /// Get a session by ID.
-    pub async fn get_session(&self, session_id: &str) -> Option<Arc<Session>> {
-        let sessions = self.sessions.read().await;
-        sessions.get(session_id).cloned()
-    }
-
-    /// Remove a session.
-    pub async fn remove_session(&self, session_id: &str) -> Option<Arc<Session>> {
-        let mut sessions = self.sessions.write().await;
-        sessions.remove(session_id)
-    }
-
-    /// Get all active sessions.
-    pub async fn active_sessions(&self) -> Vec<Arc<Session>> {
-        let sessions = self.sessions.read().await;
-        sessions.values().cloned().collect()
-    }
-
-    /// Close all sessions.
-    pub async fn close_all(&self) -> Result<(), ConnectionError> {
-        let sessions = {
-            let mut sessions = self.sessions.write().await;
-            let active: Vec<_> = sessions.drain().map(|(_, s)| s).collect();
-            active
-        };
-
-        for session in sessions {
-            session.close().await?;
-        }
-
-        Ok(())
-    }
-
-    /// Clean up idle sessions beyond timeout.
-    pub async fn cleanup_idle_sessions(&self) -> usize {
-        let sessions = self.sessions.read().await;
-        let mut to_remove = Vec::new();
-
-        for (id, session) in sessions.iter() {
-            if session.is_idle_timeout().await {
-                to_remove.push(id.clone());
-            }
-        }
-
-        drop(sessions);
-
-        let count = to_remove.len();
-        for id in to_remove {
-            if let Some(session) = self.remove_session(&id).await {
-                let _ = session.close().await;
-            }
-        }
-
-        count
-    }
-}
-
-impl std::fmt::Debug for SessionManager {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("SessionManager")
-            .field("config", &self.config)
-            .finish()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::connection::auth::{AuthResponseData, Credentials};
+    use crate::connection::auth::AuthResponseData;
 
     fn mock_server_info() -> AuthResponseData {
         mock_server_info_with_version("7.1.0")
@@ -608,47 +477,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_session_attributes() {
-        let session = Session::new(
-            "sess123".to_string(),
-            mock_server_info(),
-            SessionConfig::default(),
-        );
-
-        assert!(session.get_attribute("key1").await.is_none());
-
-        session
-            .set_attribute("key1".to_string(), "value1".to_string())
-            .await;
-        assert_eq!(
-            session.get_attribute("key1").await,
-            Some("value1".to_string())
-        );
-
-        let removed = session.remove_attribute("key1").await;
-        assert_eq!(removed, Some("value1".to_string()));
-        assert!(session.get_attribute("key1").await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_session_activity() {
-        let session = Session::new(
-            "sess123".to_string(),
-            mock_server_info(),
-            SessionConfig::default(),
-        );
-
-        session.update_activity().await;
-
-        let idle = session.idle_duration().await;
-        assert!(idle < Duration::from_millis(100));
-
-        tokio::time::sleep(Duration::from_millis(10)).await;
-        let idle = session.idle_duration().await;
-        assert!(idle >= Duration::from_millis(10));
-    }
-
-    #[tokio::test]
     async fn test_session_close() {
         let session = Session::new(
             "sess123".to_string(),
@@ -681,69 +509,6 @@ mod tests {
         // Error state should fail
         session.set_state(SessionState::Error).await;
         assert!(session.validate_ready().await.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_session_manager() {
-        let creds = Credentials::new("admin".to_string(), "secret".to_string());
-        let auth_handler = Arc::new(AuthenticationHandler::new(
-            creds,
-            "test".to_string(),
-            "1.0".to_string(),
-        ));
-        let manager = SessionManager::new(auth_handler, SessionConfig::default());
-
-        let session = Arc::new(Session::new(
-            "sess123".to_string(),
-            mock_server_info(),
-            SessionConfig::default(),
-        ));
-
-        // Register session
-        manager.register_session(session.clone()).await;
-
-        // Get session
-        let retrieved = manager.get_session("sess123").await;
-        assert!(retrieved.is_some());
-        assert_eq!(retrieved.unwrap().session_id(), "sess123");
-
-        // Active sessions
-        let active = manager.active_sessions().await;
-        assert_eq!(active.len(), 1);
-
-        // Remove session
-        let removed = manager.remove_session("sess123").await;
-        assert!(removed.is_some());
-
-        // Session no longer found
-        assert!(manager.get_session("sess123").await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_session_manager_close_all() {
-        let creds = Credentials::new("admin".to_string(), "secret".to_string());
-        let auth_handler = Arc::new(AuthenticationHandler::new(
-            creds,
-            "test".to_string(),
-            "1.0".to_string(),
-        ));
-        let manager = SessionManager::new(auth_handler, SessionConfig::default());
-
-        // Register multiple sessions
-        for i in 0..3 {
-            let session = Arc::new(Session::new(
-                format!("sess{}", i),
-                mock_server_info(),
-                SessionConfig::default(),
-            ));
-            manager.register_session(session).await;
-        }
-
-        assert_eq!(manager.active_sessions().await.len(), 3);
-
-        // Close all
-        manager.close_all().await.unwrap();
-        assert_eq!(manager.active_sessions().await.len(), 0);
     }
 
     #[test]
