+++ src/transport/native/mod.rs
@@ -10,7 +10,6 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use rustls::pki_types::CertificateDer;
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::TcpStream;
 use tokio_rustls::client::TlsStream;
@@ -21,6 +20,7 @@ use super::messages::{ColumnInfo, ResultData, ResultPayload, ResultSetHandle, Se
 use super::protocol::{
     ConnectionParams, Credentials, PreparedStatementHandle, QueryResult, TransportProtocol,
 };
+use super::tls::{FingerprintVerifier, NoVerifier};
 
 use self::attributes::{AttributeSet, AttributeValue};
 use self::constants::*;
@@ -672,115 +672,6 @@ impl NativeTcpTransport {
 impl Default for NativeTcpTransport {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-// --- TLS certificate verifiers (reused from websocket.rs) ---
-
-fn all_supported_verify_schemes() -> Vec<rustls::SignatureScheme> {
-    vec![
-        rustls::SignatureScheme::RSA_PKCS1_SHA256,
-        rustls::SignatureScheme::RSA_PKCS1_SHA384,
-        rustls::SignatureScheme::RSA_PKCS1_SHA512,
-        rustls::SignatureScheme::ECDSA_NISTP256_SHA256,
-        rustls::SignatureScheme::ECDSA_NISTP384_SHA384,
-        rustls::SignatureScheme::ECDSA_NISTP521_SHA512,
-        rustls::SignatureScheme::RSA_PSS_SHA256,
-        rustls::SignatureScheme::RSA_PSS_SHA384,
-        rustls::SignatureScheme::RSA_PSS_SHA512,
-        rustls::SignatureScheme::ED25519,
-    ]
-}
-
-#[derive(Debug)]
-struct NoVerifier;
-
-impl rustls::client::danger::ServerCertVerifier for NoVerifier {
-    fn verify_server_cert(
-        &self,
-        _end_entity: &CertificateDer<'_>,
-        _intermediates: &[CertificateDer<'_>],
-        _server_name: &rustls::pki_types::ServerName<'_>,
-        _ocsp_response: &[u8],
-        _now: rustls::pki_types::UnixTime,
-    ) -> Result<rustls::client::danger::ServerCertVerified, rustls::Error> {
-        Ok(rustls::client::danger::ServerCertVerified::assertion())
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
-        all_supported_verify_schemes()
-    }
-}
-
-#[derive(Debug)]
-struct FingerprintVerifier {
-    expected_fingerprint: String,
-}
-
-impl rustls::client::danger::ServerCertVerifier for FingerprintVerifier {
-    fn verify_server_cert(
-        &self,
-        end_entity: &CertificateDer<'_>,
-        _intermediates: &[CertificateDer<'_>],
-        _server_name: &rustls::pki_types::ServerName<'_>,
-        _ocsp_response: &[u8],
-        _now: rustls::pki_types::UnixTime,
-    ) -> Result<rustls::client::danger::ServerCertVerified, rustls::Error> {
-        use aws_lc_rs::digest;
-        let fingerprint = digest::digest(&digest::SHA256, end_entity.as_ref());
-        let actual: String = fingerprint
-            .as_ref()
-            .iter()
-            .map(|b| format!("{:02x}", b))
-            .collect();
-        if actual == self.expected_fingerprint {
-            Ok(rustls::client::danger::ServerCertVerified::assertion())
-        } else {
-            Err(rustls::Error::General(format!(
-                "Certificate fingerprint mismatch: expected {}, got {}",
-                self.expected_fingerprint, actual
-            )))
-        }
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
-        all_supported_verify_schemes()
     }
 }
